+++ starknet_py/net/account/account_deployment_result.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from starknet_py.contract import SentTransaction
-from starknet_py.utils.sync import add_sync_methods
+from starknet_py.utils.decorators import add_sync_methods
 
 
 @add_sync_methods
