+++ starknet_py/net/full_node_client.py
@@ -41,7 +41,7 @@ from starknet_py.net.schemas.rpc import (
     TypesOfTransactionsSchema,
 )
 from starknet_py.transaction_exceptions import TransactionNotReceivedError
-from starknet_py.utils.sync import add_sync_methods
+from starknet_py.utils.decorators import add_sync_methods
 
 
 @add_sync_methods
