+++ starknet_py/net/l1/messages.py
@@ -14,7 +14,7 @@ from starknet_py.net.client import Client
 from starknet_py.net.l1.contracts import StarknetL1Contract
 from starknet_py.net.l1.starknet_l1_abi import STARKNET_L1_ABI
 from starknet_py.net.models import AddressRepresentation, StarknetChainId, parse_address
-from starknet_py.utils.sync import add_sync_methods
+from starknet_py.utils.decorators import add_sync_methods
 
 
 def encode_packed(*args: int) -> bytes:
