--- starknet_py/utils/code_examples/__init__.py
@@ -1,1 +1,0 @@
-from starknet_py.utils.code_examples.code_examples import add_code_examples
