+++ starknet_py/utils/decorators/__init__.py
@@ -1,0 +1,2 @@
+from starknet_py.utils.decorators.code_examples.code_examples import add_code_examples
+from starknet_py.utils.decorators.sync.sync import add_sync_methods
