+++ starknet_py/utils/decorators/sync/sync.py
