+++ starknet_py/utils/decorators/sync/sync_test.py
@@ -2,7 +2,7 @@ import asyncio
 
 import pytest
 
-from starknet_py.utils.sync import add_sync_methods
+from starknet_py.utils.decorators import add_sync_methods
 
 
 @add_sync_methods
